Add ability to reset indexes when populating documents
