+++ backend/cosmetology-app/lambdas/python/search/handlers/manage_opensearch_indices.py
@@ -3,10 +3,7 @@ import time
 from cc_common.config import config, logger
 from cc_common.exceptions import CCInternalException
 from custom_resource_handler import CustomResourceHandler, CustomResourceResponse
-from opensearch_client import OpenSearchClient
-
-# Initial index version for new deployments
-INITIAL_INDEX_VERSION = 'v1'
+from opensearch_client import INITIAL_INDEX_VERSION, OpenSearchClient
 
 # Readiness check configuration
 # OpenSearch domains may take time to become responsive after CloudFormation reports them as created.
@@ -53,8 +50,7 @@ class OpenSearchIndexManager(CustomResourceHandler):
             index_name = f'compact_{compact}_providers_{INITIAL_INDEX_VERSION}'
             # Create alias name (e.g., compact_cosm_providers)
             alias_name = f'compact_{compact}_providers'
-            self._create_provider_index_with_alias(
-                client=client,
+            client.create_provider_index_with_alias(
                 index_name=index_name,
                 alias_name=alias_name,
                 number_of_shards=number_of_shards,
@@ -143,216 +139,6 @@ class OpenSearchIndexManager(CustomResourceHandler):
             f'({DOMAIN_READINESS_MAX_ATTEMPTS * DOMAIN_READINESS_CHECK_INTERVAL_SECONDS} seconds). '
             f'Last error: {last_exception}'
         )
-
-    def _create_provider_index_with_alias(
-        self,
-        client: OpenSearchClient,
-        index_name: str,
-        alias_name: str,
-        number_of_shards: int,
-        number_of_replicas: int,
-    ) -> None:
-        """
-        Create the provider index and alias in OpenSearch if they don't exist.
-
-        :param client: The OpenSearch client
-        :param index_name: The versioned index name (e.g., compact_cosm_providers_v1)
-        :param alias_name: The alias name (e.g., compact_cosm_providers)
-        :param number_of_shards: Number of primary shards for the index
-        :param number_of_replicas: Number of replica shards for the index
-        """
-        # Check if the alias already exists (meaning an index version is already set up)
-        if client.alias_exists(alias_name):
-            logger.info(f"Alias '{alias_name}' already exists. Skipping index and alias creation.")
-            return
-
-        # Check if the index already exists (edge case: index exists but alias doesn't)
-        if client.index_exists(index_name):
-            logger.info(f"Index '{index_name}' already exists. Creating alias only.")
-            client.create_alias(index_name, alias_name)
-            logger.info(f"Alias '{alias_name}' -> '{index_name}' created successfully.")
-            return
-
-        # Create the index with the specified configuration
-        logger.info(f"Creating index '{index_name}'...")
-        index_mapping = self._get_provider_index_mapping(number_of_shards, number_of_replicas)
-        client.create_index(index_name, index_mapping)
-        logger.info(f"Index '{index_name}' created successfully.")
-
-        # Create the alias pointing to the new index
-        logger.info(f"Creating alias '{alias_name}' -> '{index_name}'...")
-        client.create_alias(index_name, alias_name)
-        logger.info(f"Alias '{alias_name}' -> '{index_name}' created successfully.")
-
-    def _get_provider_index_mapping(self, number_of_shards: int, number_of_replicas: int) -> dict:
-        """
-        Define the index mapping for provider documents.
-
-        :param number_of_shards: Number of primary shards for the index
-        :param number_of_replicas: Number of replica shards for the index
-        :return: The index mapping dictionary
-        """
-        # Nested schema for AdverseAction
-        adverse_action_properties = {
-            'type': {'type': 'keyword'},
-            'adverseActionId': {'type': 'keyword'},
-            'compact': {'type': 'keyword'},
-            'jurisdiction': {'type': 'keyword'},
-            'providerId': {'type': 'keyword'},
-            'licenseType': {'type': 'keyword'},
-            'licenseTypeAbbreviation': {'type': 'keyword'},
-            'actionAgainst': {'type': 'keyword'},
-            'effectiveStartDate': {'type': 'date'},
-            'creationDate': {'type': 'date'},
-            'effectiveLiftDate': {'type': 'date'},
-            'dateOfUpdate': {'type': 'date'},
-            'encumbranceType': {'type': 'keyword'},
-            'clinicalPrivilegeActionCategories': {'type': 'keyword'},
-            'clinicalPrivilegeActionCategory': {'type': 'keyword'},
-            'submittingUser': {'type': 'keyword'},
-            'liftingUser': {'type': 'keyword'},
-        }
-
-        # Nested schema for Investigation
-        investigation_properties = {
-            'type': {'type': 'keyword'},
-            'investigationId': {'type': 'keyword'},
-            'compact': {'type': 'keyword'},
-            'jurisdiction': {'type': 'keyword'},
-            'licenseType': {'type': 'keyword'},
-            'status': {'type': 'keyword'},
-            'dateOfUpdate': {'type': 'date'},
-        }
-
-        # Nested schema for License
-        license_properties = {
-            'providerId': {'type': 'keyword'},
-            'type': {'type': 'keyword'},
-            'dateOfUpdate': {'type': 'date'},
-            'compact': {'type': 'keyword'},
-            'jurisdiction': {'type': 'keyword'},
-            'licenseType': {'type': 'keyword'},
-            'licenseStatusName': {'type': 'keyword'},
-            'licenseStatus': {'type': 'keyword'},
-            'jurisdictionUploadedLicenseStatus': {'type': 'keyword'},
-            'compactEligibility': {'type': 'keyword'},
-            'jurisdictionUploadedCompactEligibility': {'type': 'keyword'},
-            'licenseNumber': {'type': 'keyword'},
-            'givenName': {
-                'type': 'text',
-                'analyzer': 'custom_ascii_analyzer',
-                'fields': {'keyword': {'type': 'keyword', 'ignore_above': 256}},
-            },
-            'middleName': {
-                'type': 'text',
-                'analyzer': 'custom_ascii_analyzer',
-                'fields': {'keyword': {'type': 'keyword', 'ignore_above': 256}},
-            },
-            'familyName': {
-                'type': 'text',
-                'analyzer': 'custom_ascii_analyzer',
-                'fields': {'keyword': {'type': 'keyword', 'ignore_above': 256}},
-            },
-            'suffix': {'type': 'keyword'},
-            'dateOfIssuance': {'type': 'date'},
-            'dateOfRenewal': {'type': 'date'},
-            'dateOfExpiration': {'type': 'date'},
-            'dateOfBirth': {'type': 'date'},
-            'homeAddressStreet1': {'type': 'text'},
-            'homeAddressStreet2': {'type': 'text'},
-            'homeAddressCity': {
-                'type': 'text',
-                'analyzer': 'custom_ascii_analyzer',
-                'fields': {'keyword': {'type': 'keyword', 'ignore_above': 256}},
-            },
-            'homeAddressState': {'type': 'keyword'},
-            'homeAddressPostalCode': {'type': 'keyword'},
-            'emailAddress': {'type': 'keyword'},
-            'phoneNumber': {'type': 'keyword'},
-            'adverseActions': {'type': 'nested', 'properties': adverse_action_properties},
-            'investigations': {'type': 'nested', 'properties': investigation_properties},
-            'investigationStatus': {'type': 'keyword'},
-        }
-
-        # Nested schema for Privilege
-        privilege_properties = {
-            'type': {'type': 'keyword'},
-            'providerId': {'type': 'keyword'},
-            'compact': {'type': 'keyword'},
-            'jurisdiction': {'type': 'keyword'},
-            'licenseJurisdiction': {'type': 'keyword'},
-            'licenseType': {'type': 'keyword'},
-            'dateOfIssuance': {'type': 'date'},
-            'dateOfRenewal': {'type': 'date'},
-            'dateOfExpiration': {'type': 'date'},
-            'dateOfUpdate': {'type': 'date'},
-            'adverseActions': {'type': 'nested', 'properties': adverse_action_properties},
-            'investigations': {'type': 'nested', 'properties': investigation_properties},
-            'administratorSetStatus': {'type': 'keyword'},
-            'compactTransactionId': {'type': 'keyword'},
-            'privilegeId': {'type': 'keyword'},
-            'status': {'type': 'keyword'},
-            'investigationStatus': {'type': 'keyword'},
-        }
-
-        return {
-            'settings': {
-                'index': {
-                    'number_of_shards': number_of_shards,
-                    'number_of_replicas': number_of_replicas,
-                },
-                'analysis': {
-                    # this custom analyzer is recommended by Opensearch when you have international character
-                    # sets, and you want to support searching by their closest ASCII equivalents.
-                    # See https://docs.opensearch.org/latest/analyzers/token-filters/asciifolding/
-                    'filter': {'custom_ascii_folding': {'type': 'asciifolding', 'preserve_original': True}},
-                    'analyzer': {
-                        'custom_ascii_analyzer': {
-                            'type': 'custom',
-                            'tokenizer': 'standard',
-                            'filter': ['lowercase', 'custom_ascii_folding'],
-                        }
-                    },
-                },
-            },
-            'mappings': {
-                'properties': {
-                    # Top-level provider fields
-                    'providerId': {'type': 'keyword'},
-                    'type': {'type': 'keyword'},
-                    'dateOfUpdate': {'type': 'date'},
-                    'compact': {'type': 'keyword'},
-                    'licenseJurisdiction': {'type': 'keyword'},
-                    'licenseStatus': {'type': 'keyword'},
-                    'compactEligibility': {'type': 'keyword'},
-                    'givenName': {
-                        'type': 'text',
-                        'analyzer': 'custom_ascii_analyzer',
-                        'fields': {'keyword': {'type': 'keyword', 'ignore_above': 256}},
-                    },
-                    'middleName': {
-                        'type': 'text',
-                        'analyzer': 'custom_ascii_analyzer',
-                        'fields': {'keyword': {'type': 'keyword', 'ignore_above': 256}},
-                    },
-                    'familyName': {
-                        'type': 'text',
-                        'analyzer': 'custom_ascii_analyzer',
-                        'fields': {'keyword': {'type': 'keyword', 'ignore_above': 256}},
-                    },
-                    'suffix': {'type': 'keyword'},
-                    'dateOfExpiration': {'type': 'date'},
-                    'jurisdictionUploadedLicenseStatus': {'type': 'keyword'},
-                    'jurisdictionUploadedCompactEligibility': {'type': 'keyword'},
-                    'providerFamGivMid': {'type': 'keyword'},
-                    'providerDateOfUpdate': {'type': 'date'},
-                    'birthMonthDay': {'type': 'keyword'},
-                    # Nested arrays
-                    'licenses': {'type': 'nested', 'properties': license_properties},
-                    'privileges': {'type': 'nested', 'properties': privilege_properties},
-                }
-            },
-        }
 
 
 on_event = OpenSearchIndexManager('opensearch-index-manager')
