+++ backend/cosmetology-app/lambdas/python/search/tests/function/test_manage_opensearch_indices.py
@@ -67,6 +67,21 @@ class TestOpenSearchIndexManager(TstFunction):
         else:
             mock_client_instance.index_exists.return_value = index_exists_return_value
 
+        # Index creation lives on OpenSearchClient; delegate so alias_exists / create_index / etc. stay observable.
+        from opensearch_client import OpenSearchClient
+
+        def _real_get_mapping(number_of_shards, number_of_replicas):
+            return OpenSearchClient._get_provider_index_mapping(
+                mock_client_instance, number_of_shards, number_of_replicas
+            )
+
+        mock_client_instance._get_provider_index_mapping = _real_get_mapping
+
+        def _real_create_provider_index(*args, **kwargs):
+            return OpenSearchClient.create_provider_index_with_alias(mock_client_instance, *args, **kwargs)
+
+        mock_client_instance.create_provider_index_with_alias.side_effect = _real_create_provider_index
+
         return mock_client_instance
 
     @patch('handlers.manage_opensearch_indices.OpenSearchClient')
